refactor: move config into template
